EXPERIMENTO/PRUEBA GROC/requirements.txt
langchain>=0.1.0
langchain-groq>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0